requirements.txt
nest-asyncio
haystack-ai
fastapi